+++ src/ChannelServer/ChannelServer.h
@@ -19,6 +19,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #define CHANNELSERVER_H
 
 #include "AbstractServer.h"
+#include "Types.h"
 #include <string>
 
 using std::string;
@@ -44,25 +45,25 @@ public:
 	char getWorld() { return world; }
 	void setWorld(char id) { world = id; }
 	void setWorldIp(const string &ip) { world_ip = ip; }
-	void setWorldPort(short port) { world_port = port; }
+	void setWorldPort(int16_t port) { world_port = port; }
 	WorldServerConnectPlayer * getWorldPlayer() { return worldPlayer; }
-	int getChannel() { return channel; }
-	void setChannel(int channel) { this->channel = channel; }
-	void setPort(short port) { this->port = port; }
-	int getOnlineId() { return 20000 + (int) world * 100 + channel; }
+	uint16_t getChannel() { return channel; }
+	void setChannel(uint16_t channel) { this->channel = channel; }
+	void setPort(int16_t port) { this->port = port; }
+	int32_t getOnlineId() { return 20000 + (int32_t) world * 100 + channel; }
 	string getScrollingHeader() { return scrollingHeader; }
 	void setScrollingHeader(const string &message);
 	unsigned char getMaxMultiLevel() const { return maxMultiLevel; }
 	void setMaxMultiLevel(unsigned char level) { maxMultiLevel = level; }
 	// Server rates
-	void setExprate(int exprate) { this->exprate = exprate; }
-	void setQuestExprate(int questexprate) { this->questexprate = questexprate; }
-	void setMesorate (int mesorate) { this->mesorate = mesorate; }
-	void setDroprate (int droprate) { this->droprate = droprate; }
-	int getExprate() { return exprate; }
-	int getQuestExprate() { return questexprate; }
-	int getMesorate() { return mesorate; }
-	int getDroprate() { return droprate; }
+	void setExprate(int32_t exprate) { this->exprate = exprate; }
+	void setQuestExprate(int32_t questexprate) { this->questexprate = questexprate; }
+	void setMesorate (int32_t mesorate) { this->mesorate = mesorate; }
+	void setDroprate (int32_t droprate) { this->droprate = droprate; }
+	int32_t getExprate() { return exprate; }
+	int32_t getQuestExprate() { return questexprate; }
+	int32_t getMesorate() { return mesorate; }
+	int32_t getDroprate() { return droprate; }
 private:
 	ChannelServer() {};
 	ChannelServer(const ChannelServer&);
@@ -72,20 +73,20 @@ private:
 	WorldServerConnectPlayer *worldPlayer;
 
 	string login_ip;
-	short login_inter_port;
-	short port;
+	uint16_t login_inter_port;
+	uint16_t port;
 	char world;
 	string world_ip;
-	short world_port;
-	int channel;
+	uint16_t world_port;
+	uint16_t channel;
 	string external_ip;
 	string scrollingHeader;
 	unsigned char maxMultiLevel;
 	// Server rates
-	int exprate;
-	int questexprate;
-	int mesorate;
-	int droprate;
+	int32_t exprate;
+	int32_t questexprate;
+	int32_t mesorate;
+	int32_t droprate;
 };
 
 #endif
